requirements.txt
streamlit>=1.28.0
nltk>=3.8.1
PyPDF2>=3.0.0
python-docx>=1.0.0
pycryptodome>=3.19.0
